optimize: repackage spring-boot-starter (#2601)
