+++ seata-spring-boot-starter/src/main/java/io/seata/spring/boot/autoconfigure/properties/client/RmProperties.java
@@ -13,7 +13,7 @@
  *  See the License for the specific language governing permissions and
  *  limitations under the License.
  */
-package io.seata.spring.boot.autoconfigure.properties.file;
+package io.seata.spring.boot.autoconfigure.properties.client;
 
 import org.springframework.boot.context.properties.ConfigurationProperties;
 import org.springframework.stereotype.Component;
