+++ seata-spring-boot-starter/src/main/java/io/seata/spring/boot/autoconfigure/properties/client/ServiceProperties.java
@@ -13,7 +13,7 @@
  *  See the License for the specific language governing permissions and
  *  limitations under the License.
  */
-package io.seata.spring.boot.autoconfigure.properties.file;
+package io.seata.spring.boot.autoconfigure.properties.client;
 
 import java.util.HashMap;
 import java.util.Map;
